Extract breadcrumbs
